+++ pandora_console/extras/pandoradb_migrate_6.0_to_7.0.mysql.sql
@@ -4202,7 +4202,7 @@ CREATE TABLE IF NOT EXISTS `tncm_agent_data` (
 CREATE TABLE IF NOT EXISTS `tncm_queue` (
 	`id` SERIAL,
     `id_agent` INT(10) UNSIGNED NOT NULL,
-    `id_script` INT(10) UNSIGNED NOT NULL,
+    `id_script` BIGINT(20) UNSIGNED NOT NULL,
 	`utimestamp` INT UNSIGNED NOT NULL,
     FOREIGN KEY (`id_agent`) REFERENCES `tagente`(`id_agente`) ON UPDATE CASCADE ON DELETE CASCADE,
 	FOREIGN KEY (`id_script`) REFERENCES `tncm_script`(`id`) ON UPDATE CASCADE ON DELETE CASCADE
